Session and authentication (Login example)

+++ src/Controller/BookmarksController.php
@@ -18,7 +18,7 @@ class BookmarksController extends AppController
     //         'Articles.title' => 'asc'
     //     ]
     // ];
-
+    public $user = null;
 
     public function initialize()
     {
@@ -26,6 +26,9 @@ class BookmarksController extends AppController
         
         // Set the layout
         // $this->layout = 'frontend';
+
+        $this->user= $this->Auth->user('id');
+
         $this->viewBuilder()->layout('frontend_bookmark');
     }
     /**
@@ -100,23 +103,30 @@ class BookmarksController extends AppController
      */
     public function edit($id = null)
     {
-        $bookmark = $this->Bookmarks->get($id, [
-            'contain' => ['Tags']
-        ]);
-        if ($this->request->is(['patch', 'post', 'put'])) {
-            $bookmark = $this->Bookmarks->patchEntity($bookmark, $this->request->data);
-            if ($this->Bookmarks->save($bookmark)) {
-                $this->Flash->success(__('The bookmark has been saved.'));
+        if($this->isAuthorized($this->user,$this->request->session()->read('Auth.User.username')))
+        { 
+            $bookmark = $this->Bookmarks->get($id, [
+                'contain' => ['Tags']
+            ]);
+            if ($this->request->is(['patch', 'post', 'put'])) {
+                $bookmark = $this->Bookmarks->patchEntity($bookmark, $this->request->data);
+                if ($this->Bookmarks->save($bookmark)) {
+                    $this->Flash->success(__('The bookmark has been saved.'));
 
-                return $this->redirect(['action' => 'index']);
-            } else {
-                $this->Flash->error(__('The bookmark could not be saved. Please, try again.'));
+                    return $this->redirect(['action' => 'index']);
+                } else {
+                    $this->Flash->error(__('The bookmark could not be saved. Please, try again.'));
+                }
             }
+            $users = $this->Bookmarks->Users->find('list', ['limit' => 200]);
+            $tags = $this->Bookmarks->Tags->find('list', ['limit' => 200]);
+            $this->set(compact('bookmark', 'users', 'tags'));
+            $this->set('_serialize', ['bookmark']);
         }
-        $users = $this->Bookmarks->Users->find('list', ['limit' => 200]);
-        $tags = $this->Bookmarks->Tags->find('list', ['limit' => 200]);
-        $this->set(compact('bookmark', 'users', 'tags'));
-        $this->set('_serialize', ['bookmark']);
+        else{
+            $this->Flash->error(_('Unauthorized user'));
+            return $this->redirect(['action' => 'index']);
+        }
     }
 
     /**
@@ -128,15 +138,23 @@ class BookmarksController extends AppController
      */
     public function delete($id = null)
     {
-        $this->request->allowMethod(['post', 'delete']);
-        $bookmark = $this->Bookmarks->get($id);
-        if ($this->Bookmarks->delete($bookmark)) {
-            $this->Flash->success(__('The bookmark has been deleted.'));
-        } else {
-            $this->Flash->error(__('The bookmark could not be deleted. Please, try again.'));
-        }
+        if($this->isAuthorized($this->user,$this->request->session()->read('Auth.User.username')))
+        {
+            $this->Flash->success(__('authorized'));
+            $this->request->allowMethod(['post', 'delete']);
+            $bookmark = $this->Bookmarks->get($id);
+            if ($this->Bookmarks->delete($bookmark)) {
+                $this->Flash->success(__('The bookmark has been deleted.'));
+            } else {
+                $this->Flash->error(__('The bookmark could not be deleted. Please, try again.'));
+            }
 
-        return $this->redirect(['action' => 'index']);
+            return $this->redirect(['action' => 'index']);
+        }
+        else{
+            $this->Flash->error(_('Unauthorized user'));
+            return $this->redirect(['action' => 'index']);
+        }
     }
 
     /**
@@ -150,4 +168,22 @@ class BookmarksController extends AppController
         $this->set(compact('bookmark'));
         $this->set('_serialize', ['bookmark']);
     }
+
+    public function isAuthorized($user,$username)
+    {
+        // All registered users can add articles
+        if ($this->request->action === 'add') {
+            return true;
+        }
+
+        // The owner of an article can edit and delete it
+        if (in_array($this->request->action, ['edit', 'delete'])) {
+            $bookmarkId = (int)$this->request->params['pass'][0];            
+            if ($this->Bookmarks->isOwnedBy($bookmarkId, $this->user)) {
+                return true;
+            }
+        }
+
+        return parent::isAuthorized($user,$username);
+    }                   
 }
